Re #18421. Move API::SpectrumInfo getter into ExperimentInfo.
